refactor: remove comparison-utils and reorganize utility functions
- Deleted the comparison-utils and moved relevant functions to new utility files and the use-comparison composable
- Refactored use-comparison composable to utilize the utility functions
- Added new field-utils for field-related checks (e.g., isPrimaryKey, isRelational)
- Added new get-revision-fields utility to handle revision field filtering logic
- Added new get-version-display-name utility for version display name formatting
