+++ app/src/views/private/components/revision-item.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import { getRevisionFields } from '@/modules/content/comparison-utils';
 import { useFieldsStore } from '@/stores/fields';
 import { Revision } from '@/types/revisions';
+import { getRevisionFields } from '@/utils/get-revision-fields';
 import { userName } from '@/utils/user-name';
 import { format } from 'date-fns';
 import { computed } from 'vue';
